Add: Image share option for dive plans (#18)
Right now this serves as an initial starting point, a sort of MVP implementation. It would be good to offer more information in shared images in the future; like graphs, additional general information (max depth, average depth) and more. Also the way in which compose is used to render images is experimental at best.

Implements: #1