Refactor: Typeface, weights and sizes used for body and label styles (#27)
This makes the font sizes a bit more consistent across the app. This also switches the main font to Roboto, it used to be Source Sans (but was never really applied, so in reality it was the default Compose font). Roboto is a bit better suited for condense displayed information.