Change: Improve empty cylinder alert and modify warning color (#28)
The cylinder alert now only shows "empty" if both the baseline usage and out-of-air situation leads to an empty cylinder. If only the out-of-air situation leads to an empty cylinder it is shown as 0 with a alert symbol to the left of it.

Also the alert color for warnings has changed from a yellow-greenish color to a more orange tint.