Fix: Show both normal and deco PPO2 MOD
When selecting a cylinder only the deco MOD was being shown, while it is not yet clear what this cylinder is going to be used for. So instead as kinda of a quick fix, the app now shows both MODs.

This also improves the gas properties display a bit, especially on smaller screens.

Fixes: #23