Release: 1.0.7-beta (#66)
This release makes a major change to the way the gas plan is presented, the pie chart is gone and a new bar chart takes its place. All details related to this change can be found in this PR: #65. Other then these changes no things have changed since release 1.0.6-beta.